Create SharedDigit.java
Write a method named hasShared Digit with two parameters of type int.
Each number should be within the range of 10 (inclusive) - 99 (inclusive). If one of the numbers
is not within the range, the method should return false.
The method should return true if there is a digit that appears in both numbers, such as 2 in 12
and 23; otherwise, the method should return false.